Seperated os dependent codes
