+++ CMakeLists.txt
@@ -3,6 +3,7 @@ set(CMAKE_CXX_STANDARD 11)
 
 project(intothedungeon)
 
-#add_subdirectory(core)
+add_subdirectory(core)
 #add_subdirectory(main)
 add_subdirectory(test)
+add_subdirectory(linux)
