--- core/Point.cpp
@@ -1,19 +1,0 @@
-#include "Point.h"
-
-
-void Point::gotoxy(int x, int y)
-{
-	COORD Pos = {x, y}; 
-	SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), Pos);
-}
-
-void Point::gotoxy(Point pos)
-{
-	gotoxy(pos.getX(), pos.getY());
-}
-
-Point Point::GetScrPosFromCurPos(Point &pos) // 배열좌표를 화면좌표로 가져온다.
-{
-	pos.setX(2 * pos.m_x + 2);
-	pos.setY(pos.m_y + 1);
-}
